+++ fractal_server/app/routes/api/v2/history.py
@@ -123,19 +123,17 @@ async def get_workflow_tasks_statuses(
                 .where(
                     HistoryImageCache.latest_history_unit_id == HistoryUnit.id
                 )
-                .where(HistoryUnit.status == target_status.value)
+                .where(HistoryUnit.status == target_status)
             )
             res = await db.execute(stm)
             num_images = res.scalar()
-            response[wftask.id][
-                f"num_{target_status.value}_images"
-            ] = num_images
+            response[wftask.id][f"num_{target_status}_images"] = num_images
 
     new_response = deepcopy(response)
     for key, value in response.items():
         if value is not None:
             num_total_images = sum(
-                value[f"num_{target_status.value}_images"]
+                value[f"num_{target_status}_images"]
                 for target_status in HistoryUnitStatus
             )
             if num_total_images > value["num_available_images"]: