+++ fractal_server/app/schemas/v2/history.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from enum import Enum
+from enum import StrEnum
 from typing import Any
 from typing import Optional
 
@@ -10,7 +10,7 @@ from pydantic import field_serializer
 from ....images import SingleImage
 
 
-class HistoryUnitStatus(str, Enum):
+class HistoryUnitStatus(StrEnum):
     """
     Available status for images
 
@@ -25,7 +25,7 @@ class HistoryUnitStatus(str, Enum):
     FAILED = "failed"
 
 
-class HistoryUnitStatusQuery(str, Enum):
+class HistoryUnitStatusQuery(StrEnum):
 
     SUBMITTED = "submitted"
     DONE = "done"