+++ fractal_server/tasks/v2/ssh/collect.py
@@ -166,7 +166,7 @@ def collect_ssh(
                     script_dir_remote=script_dir_remote,
                     prefix=(
                         f"{int(time.time())}_"
-                        f"{TaskGroupActivityActionV2.COLLECT.value}"
+                        f"{TaskGroupActivityActionV2.COLLECT}"
                     ),
                     fractal_ssh=fractal_ssh,
                     logger_name=LOGGER_NAME,